addon/rootfs/app/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
jinja2==3.1.3
python-multipart==0.0.6

# EnOcean Communication
pyserial==3.5

# MQTT
paho-mqtt==1.6.1

# XML/YAML Processing
lxml==5.1.0
pyyaml==6.0.1

# Database
aiosqlite==0.19.0

# Async support
aiofiles==23.2.1
aiohttp==3.9.1
